+++ eqcorrscan/core/match_filter/party.py
@@ -37,6 +37,7 @@ from eqcorrscan.core.match_filter.matched_filter import (
     MatchFilterError, _group_process)
 from eqcorrscan.core.match_filter.template import Template
 from eqcorrscan.core.match_filter.family import Family
+from eqcorrscan.core.match_filter.detection import write_detections
 from eqcorrscan.core.match_filter.helpers import (
     _total_microsec, temporary_directory, _safemembers, _templates_match)
 
@@ -639,8 +640,8 @@ class Party(object):
                 raise MatchFilterError(
                     'Will not overwrite existing file: %s' % filename)
             for family in self.families:
-                for detection in family.detections:
-                    detection.write(fname=filename, append=True)
+                write_detections(fname=filename, detections=family.detections,
+                                 mode="a")
         elif format.lower() == 'tar':
             if os.path.exists(filename):
                 raise IOError('Will not overwrite existing file: %s'
@@ -747,7 +748,7 @@ class Party(object):
                  horizontal_chans=['E', 'N', '1', '2'], vertical_chans=['Z'],
                  cores=1, interpolate=False, plot=False, parallel=True,
                  overlap='calculate', process_cores=None,
-                 relative_magnitudes=False, **kwargs):
+                 ignore_bad_data=False, relative_magnitudes=False, **kwargs):
         """
         Compute picks based on cross-correlation alignment.
 
@@ -804,7 +805,11 @@ class Party(object):
             information. Keyword arguments `noise_window`, `signal_window` and
             `min_snr` can be passed as additional keyword arguments to pass
             through to `eqcorrscan.utils.mag_calc.relative_magnitude`.
-
+        :type ignore_bad_data: bool
+        :param ignore_bad_data:
+            If False (default), errors will be raised if data are excessively
+            gappy or are mostly zeros. If True then no error will be raised,
+            but an empty trace will be returned (and not used in detection).
 
         :returns:
             Catalog of events with picks.  No origin information is included.
@@ -885,7 +890,8 @@ class Party(object):
                 processed_streams = _group_process(
                     template_group=group, cores=process_cores,
                     parallel=parallel, stream=stream.copy(), daylong=False,
-                    ignore_length=False, overlap=lap)
+                    ignore_length=False, overlap=lap,
+                    ignore_bad_data=ignore_bad_data)
                 processed_stream = Stream()
                 for p in processed_streams:
                     processed_stream += p