+++ CHANGES.md
@@ -57,6 +57,9 @@
 * Added `relative_magnitudes` argument to 
   `eqcorrscan.core.match_filter.party.Party.lag_calc` to provide an in-flow
   way to compute relative magnitudes for detected events.
+* Fix #298 where the header was repeated in detection csv files. Also added
+  a `write_detections` function to `eqcorrscan.core.match_filter.detection`
+  to streamline writing detections.
 
 ## 0.3.3
 * Make test-script more stable.